make build simpler and cache the dependencies
One of the biggest issues people have had with snakegit has been
the two stage build (deps then build).  These changes combine all
of the work into the build step.  It also allows for downloading
dependencies.  It still caches the dependencies, but it uses the
cache format.  Also, as part of this, the pypi server has been
configured correctly to proxy to the central pypi server for packages
that are not currently available on the ns server